Improve typing.
