+++ src/components/text-pane/text-pane.scss
