+++ src/components/text-pane/text-section.scss
