+++ src/managers/text_feedback_manager.ts
@@ -8,9 +8,10 @@ import { Descendant } from "@concord-consortium/slate-editor";
 import { datasetExists, getCaseValues, getSelectedCasesFrom } from "../lib/codap-helper";
 import codapInterface, { CODAP_Notification } from "../lib/CodapInterface";
 import { featureStore } from "../stores/feature_store";
+import { ITextSectionText } from "../stores/store_types_and_constants";
 import { targetStore } from "../stores/target_store";
 import { testingStore } from "../stores/testing_store";
-import { ITextSectionText, textStore } from "../stores/text_store";
+import { textStore } from "../stores/text_store";
 import { trainingStore } from "../stores/training_store";
 import { uiStore } from "../stores/ui_store";
 import { APIRequest, GetCaseByIDResponse, GetSelectionListResponse } from "../types/codap-api-types";
